# Speed up dask.bag.Bag.random_sample