--- a/dask/bag/core.py
+++ b/dask/bag/core.py
@@ -673,11 +673,11 @@ class Bag(DaskMethodsMixin):
         Examples
         --------
         >>> import dask.bag as db
-        >>> b = db.from_sequence(range(5))
-        >>> list(b.random_sample(0.5, 43))
-        [0, 1, 4]
-        >>> list(b.random_sample(0.5, 43))
-        [0, 1, 4]
+        >>> b = db.from_sequence(range(10))
+        >>> b.random_sample(0.5, 43).compute()
+        [0, 1, 3, 4, 7, 9]
+        >>> b.random_sample(0.5, 43).compute()
+        [0, 1, 3, 4, 7, 9]
         """
         if not 0 <= prob <= 1:
             raise ValueError("prob must be a number in the interval [0, 1]")