Closes #10351
```python
%timeit random_state_data_python(1, 0)
%timeit random_state_data_python(10, 0)
%timeit random_state_data_python(100, 0)
%timeit random_state_data_python(1000, 0)
```
Before:
```
265 µs ± 1.72 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
2.64 ms ± 16.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
26.5 ms ± 146 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
272 ms ± 1.06 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
After:
```
131 µs ± 1.47 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
234 µs ± 1.21 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1.26 ms ± 9.68 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
20.4 ms ± 137 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```